:recycle: Add applyToken file util
